backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
transformers==4.36.0
torch==2.1.0
accelerate==0.24.0
sentencepiece==0.1.99
httpx==0.25.2
huggingface-hub==0.20.3
